Consolidate persistence writes onto their resource mixins
Feature mixins (progress sync, restart, rollback, scale) were writing
publish-record status/ext and device-binding rows directly via
_publish_service, scattering each table's writes across the package. Enforce
single-writer per resource: feature mixins orchestrate, resource mixins (and
the PublishExtState plumbing collaborator) persist.

- PublishExtMixin gains _record_sync_success (clear retry flag + atomic
  source->target advance with ext — the persistence half of
  _handle_sync_success) and takes over _mark_previous_publish_superseded from
  StageStatusMixin (it is a publish-record status+ext write); both route
  through the ext_state plumbing instead of raw
  _publish_service.update_publish_status_with_ext.
- DeviceBindingMixin takes over _update_binding_on_success from
  ProgressSyncMixin and gains _release_binding (extracted from rollback's
  _destroy_bot_by_stage) — all device-binding writes now live in one mixin.
- StageStatusMixin shrinks to the pure stage-from-status derivations.
- retry() and the process() double-submit CAS route through PublishExtState
  (new advance_status for the status-only CAS) instead of raw
  _publish_service calls, making ext_state the sole owner of publish-record
  writes and DeviceBindingMixin the sole owner of binding writes.

No behavior change; no test edits needed (cross-mixin calls resolve on the
same instance, so existing interception seams still hold).
